Moving vendor app to separate directory
