+++ tests/apps/vendor_app/test_vendor_management_view.py
@@ -2,7 +2,7 @@ import pytest
 from pytest_mock import MockerFixture
 import requests
 
-from plugins.vendor.vendors import VendorManagementView
+from plugins.vendor_app.vendors import VendorManagementView
 from tests.mocks import mock_okapi_variable, MockFOLIOClient
